Pass only relevant slices of WCIF
